Move test fixtures to tests/conftest.py
